+++ src/main/java/subway/repository/StationRepository.java
@@ -1,9 +1,10 @@
-package subway.domain;
+package subway.repository;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
+import subway.domain.station.Station;
 
 public class StationRepository {
     private static final List<Station> stations = new ArrayList<>();
